+++ .gitignore
@@ -2,3 +2,4 @@
 **/*.egg-info
 **/*~
 **/__pycache__
+**/.ipynb_checkpoints
